+++ server/modules/selva/module/selva_object.h
@@ -27,7 +27,7 @@ int SelvaObject_GetDouble(struct SelvaObject *obj, const struct RedisModuleStrin
 int SelvaObject_GetLongLong(struct SelvaObject *obj, const struct RedisModuleString *key_name, long long *out);
 int SelvaObject_GetStr(struct SelvaObject *obj, const struct RedisModuleString *key_name, struct RedisModuleString **out);
 int SelvaObject_SetDouble(struct SelvaObject *obj, const struct RedisModuleString *key_name, double value);
-int SelvaObject_SetLongLong(struct SelvaObject *obj, const struct RedisModuleString *key_name, double value);
+int SelvaObject_SetLongLong(struct SelvaObject *obj, const struct RedisModuleString *key_name, long long value);
 int SelvaObject_SetStr(struct SelvaObject *obj, const struct RedisModuleString *key_name, struct RedisModuleString *value);
 int SelvaObject_AddSet(struct SelvaObject *obj, const struct RedisModuleString *key_name, struct RedisModuleString *value);
 int SelvaObject_AddArray(struct SelvaObject *obj, const struct RedisModuleString *key_name, enum SelvaObjectType subtype, void *p);
